+++ test/cljs/lambdacd/testdata.cljs
@@ -1,4 +1,9 @@
-(ns lambdacd.testdata)
+(ns lambdacd.testdata
+  (:require [lambdacd.time :as time]
+            [cljs-time.core :as t]))
+
+(def time-start (t/now))
+(def time-after-ten-sec (t/plus time-start (t/seconds 10)))
 
 (def some-build-step
   {:name "some-step"
@@ -6,7 +11,9 @@
    :step-id [1 2 3]
    :children []
    :result {:status "success"
-            :out "hello world"}})
+            :out "hello world"
+            :first-updated-at (time/unparse-time time-start)
+            :most-recent-update-at (time/unparse-time time-after-ten-sec)}})
 
 (defn with-name [step name]
   (assoc step :name name))
@@ -22,3 +29,9 @@
 
 (defn with-children [step children]
   (assoc step :children children))
+
+(defn with-most-recent-update [step ts]
+  (assoc-in step [:result :most-recent-update-at] ts))
+
+(defn with-first-update-at [step ts]
+  (assoc-in step [:result :first-updated-at] ts))